Move parameters out into separate classes.
* FQ
* Hashed
* N
* ProductSort
* Q
* QT
* Rows
* Sort
