+++ src/main/java/com/github/thesench/solr/dsl/stream/expr/params/ProductSort.java
@@ -7,4 +7,12 @@ public class ProductSort extends StreamExpressionNamedParameter {
     public ProductSort(String sorts) {
         super("productSort", sorts);
     }
+
+    public static ProductSort productSort(SortFields sortFields) {
+        return new ProductSort(sortFields.toString());
+    }
+
+    public static ProductSort productSort(String sorts) {
+        return new ProductSort(sorts);
+    }
 }
