+++ src/main/java/com/github/thesench/solr/dsl/stream/expr/params/Sort.java
@@ -9,4 +9,16 @@ public class Sort extends StreamExpressionNamedParameter implements SearchParame
     public Sort(String sort) {
         super(CommonParams.SORT, sort);
     }
+
+    public static Sort sort(SortFields sortField) {
+        return new Sort(sortField.toString());
+    }
+
+    public static Sort sort(String sort) {
+        return new Sort(sort);
+    }
+
+    public static Sort sort(String... sorts) {
+        return new Sort(String.join(",", sorts));
+    }
 }
