+++ src/main/java/com/github/thesench/solr/dsl/stream/expr/params/StreamParameters.java
@@ -25,55 +25,11 @@ public class StreamParameters {
         return new FL(fields);
     }
 
-    public static FQ fq(String filterQuery) {
-        return new FQ(filterQuery);
-    }
-
     public static Reducer group(Sort sort, N n) {
         return new Reducer("group", sort, n);
     }
 
-    public static Hashed hashed(StreamExpression stream) {
-        return new Hashed(stream);
-    }
-
-    public static N n(int n) {
-        return new N(n);
-    }
-
-    public static ProductSort productSort(SortFields sortFields) {
-        return new ProductSort(sortFields.toString());
-    }
-
-    public static ProductSort productSort(String sorts) {
-        return new ProductSort(sorts);
-    }
-
-    public static Q q(String query) {
-        return new Q(query);
-    }
-
-    public static QT qt(RequestHandler requestHandler) {
-        return new QT(requestHandler);
-    }
-
     public static RawSearchParameter raw(String name, String value) {
         return new RawSearchParameter(name, value);
-    }
-
-    public static Rows rows(int numRows) {
-        return new Rows(numRows);
-    }
-
-    public static Sort sort(SortFields sortField) {
-        return new Sort(sortField.toString());
-    }
-
-    public static Sort sort(String sort) {
-        return new Sort(sort);
-    }
-
-    public static Sort sort(String... sorts) {
-        return new Sort(String.join(",", sorts));
     }
 }
